pyproject.toml
[project]
name = "plugnbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cachetools>=6.2.0",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.116.1",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "huggingface-hub[hf-xet]>=0.34.4",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-groq>=0.3.7",
    "langchain-huggingface>=0.3.1",
    "pydantic>=2.11.7",
    "pypdf>=6.0.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "uvicorn>=0.35.0",
]
